+++ ChangeLog.md
@@ -1,4 +1,31 @@
 # Changelog
+# All notable changes to this project will be documented in this file.
+
+## v1.4.1 r3
+* Working EAI Sorting mode. Enable it checking the "EAI Sorting" option (This option will be remembered in the future)
+* Re worked enabling and disabling assets buttons based on selected assets.
+* No more check boxes in the item list. Select the assets like any icon in File Explorer then Disable or Enable the assets.
+* Fixed broken json after saving a surface asset in the editor.
+* Improved asset loading in the list.
+
+## v1.4.1 r1
+* Modified selection of assets
+* Fixed enabled and disabled assets mixed in list
+
+## v1.4.0 r3
+* Added ""loading"" indicator
+
+## v1.4.0 r2
+* Fixed invalid json when saving invalid loaded json on editor
+
+## V1.4.0
+* Added Custom Assets Installer
+
+## v1.3.9
+* Added button icons
+* Added more mod info in the InfoBar
+* Fixed InfoBar hanging on resize
+
 ## v1.3.8
 * Changed UiPriority and Draw Order controls